requirements.txt
fastapi
uvicorn
sqlalchemy
passlib
pymysql
python-jose
cryptography
python-dotenv